backend/requirements.txt
# ── Web framework ──────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12

# ── LLM ─────────────────────────────────────────
anthropic==0.39.0

# ── NLI e ML ────────────────────────────────────
torch==2.5.1
transformers==4.46.2
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==2.1.3

# ── Interpretability ────────────────────────────
captum==0.7.0